# Optimize Rotations::rotate() more