--- a/cpp/include/rotations.hpp
+++ b/cpp/include/rotations.hpp
@@ -2,7 +2,6 @@
 #ifndef OPENCUBES_ROTATIONS_HPP
 #define OPENCUBES_ROTATIONS_HPP
 #include <array>
-#include <vector>
 
 #include "cube.hpp"
 
@@ -15,6 +14,6 @@ struct Rotations {
         {1, 2, 0, -1, 1, -1}, {1, 2, 0, 1, -1, -1},  {1, 2, 0, 1, 1, 1},   {2, 0, 1, -1, -1, 1},  {2, 0, 1, -1, 1, -1}, {2, 0, 1, 1, -1, -1},
         {2, 0, 1, 1, 1, 1},   {2, 1, 0, -1, -1, -1}, {2, 1, 0, -1, 1, 1},  {2, 1, 0, 1, -1, 1},   {2, 1, 0, 1, 1, -1},
     };
-    static std::pair<XYZ, std::vector<XYZ>> rotate(int i, XYZ shape, const Cube &orig);
+    static std::pair<XYZ, bool> rotate(int i, XYZ shape, const Cube &orig, Cube &dest);
 };
 #endif