Bunch of small optimizations I made:
- Hoist nearly all allocations out from the hot loop of checking the rotations.
- Add out argument as `Rotations::rotate(...,Cube&)` and make it return `std::pair<XYZ,bool>` the second==true if rotation is valid. This avoids allocating memory every time in rotate().
- Update/Fix test_rotations.cpp to work with changed rotate()
- Cube::empty_from() added to clear cube and reserve enough space.
- clang-format the sources
- Misc fixups to the cubes code. 

this:
N = 12 || generating new cubes from 2522522 base cubes. 66 sets by shape for N=12
converting to vector
starting 2 threads
  done took 98.79 s [      0, 1261261]
  done took 114.17 s [1261261, 2522522]
  num cubes: 18598427
./build/cubes -n 12 -t 2  245.05s user 1.88s system 180% cpu 2:16.76 total
versus:
N = 12 || generating new cubes from 2522522 base cubes.
66 sets by shape for N=12
converting to vector
starting 2 threads
  done took 119.84 s [      0, 1261261]
  done took 139.10 s [1261261, 2522522]
  num cubes: 18598427
./build/cubes -n 12 -t 2  297.87s user 2.04s system 180% cpu 2:46.07 total


N=12 is biggest that I can test and most of time is still spent sorting the Cube rotations...
The Cube::empty_from() might temporal thing, since I have plans to make the Cube constant sized at point of construction.
There is no real need for dynamically growing vector as the Cube size is known.

Objections?